fashion_medallion/requirements.txt
pyspark==3.5.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
duckdb==1.1.3
streamlit==1.39.0
altair==5.4.1
pandas==2.2.3
pytest==8.3.3
httpx==0.27.2
